requirements.txt
setuptools>=65.5.0
wheel>=0.38.0
Flask==3.0.0
Flask-CORS==4.0.0
opencv-python==4.8.1.78
librosa==0.10.1
numpy==1.24.3
moviepy==1.0.3
openai-whisper==20231117
anthropic==0.7.8
Werkzeug==3.0.1
ffmpeg-python==0.2.0
torch>=2.0.0
torchaudio>=2.0.0